# apps/api/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Livecampus FrameworkJS",
    "description": "A simple api to request data for an ecommerce website",
    "version": "1.0.0",
    "contact": {
      "name": "Theo Posty",
      "email": "thso@posty.fr",
      "url": "https://theo.posty.fr"
    }
  },
  "produces": ["application/json"],
  "servers": [{ "url": "http://localhost:8000/api/v1/" }],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "NewUser": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": {
            "type": "string",
            "default": "example@email.fr"
          },
          "password": {
            "type": "string",
            "default": "Azerty1!"
          }
        }
      },
...
      "AuthUser": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "default": "example@email.fr"
          },
          "role": {
            "type": "string",
            "enum": ["admin", "user"],
            "default": "user"
          },
          "token": {
            "type": "string",
            "default": "some_token_string"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/auth/register": {
      "post": {
        "tags": ["AUTH"],
        "summary": "Register a new user. Duplicates are not allowed",
        "description": "Register a new user. Duplicates are not allowed",
        "operationId": "registerUser",
        "requestBody": {
          "description": "User to add",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewUser"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AuthUser"
                    },
                    "error": {
                      "type": "string",
                      "nullable": true,
                      "default": null
                    },
                    "statusCode": {
                      "type": "integer",
                      "format": "int8",
                      "default": 201
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "default": null
                    },
                    "error": {
                      "type": "string",
                      "default": "error message"
                    },
                    "statusCode": {
                      "type": "integer",
                      "format": "int8",
                      "default": 400
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["AUTH"],
        "summary": "Login as a user",
        "description": "Login as a user",
        "operationId": "loginUser",
        "requestBody": {
          "description": "User to login into",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewUser"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AuthUser"
                    },
                    "error": {
                      "type": "string",
                      "nullable": true,
                      "default": null
                    },
                    "statusCode": {
                      "type": "integer",
                      "format": "int8",
                      "default": 201
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "default": null
                    },
                    "error": {
                      "type": "string",
                      "default": "error message"
                    },
                    "statusCode": {
                      "type": "integer",
                      "format": "int8",
                      "default": 400
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": ["AUTH"],
        "summary": "Check and get current user",
        "description": "Check and get current user",
        "operationId": "me",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/AuthUser"
                    },
                    "error": {
                      "type": "string",
                      "nullable": true,
                      "default": null
                    },
                    "statusCode": {
                      "type": "integer",
                      "format": "int8",
                      "default": 201
                    }
                  }
                }
              }
            }
          },
          "default": {
            "description": "unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "nullable": true,
                      "default": null
                    },
                    "error": {
                      "type": "string",
                      "default": "error message"
                    },
                    "statusCode": {
                      "type": "integer",
                      "format": "int8",
                      "default": 400
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}